requirements.txt
numpy==1.23.5
scikit-image==0.19.3
opencv-python-headless==4.7.0.72
seaborn==0.12.2
scikit-learn==1.2.2
joblib==1.3.2
matplotlib==3.7.1
Pillow
streamlit